Fixed group creation bug
I fixed a bug related to p2p group creation. Previously, I was unable to see whether a group was formed (event though the connection itself was) or ascertain whether the Android device I was using was the host or the client. Apparently, connecting to legacy devices requires a createGroup function rather than a regular connect.